T1 is white letter and T2 is y/n response to an X
Modify stimulus processing to use the Wong & Chang stimuli and allow for
stimuli to be white or black. Adjust the rest of the code to collect a
broader range of responses to T1 and a y/n to T2. Various other changes to
support this. Still need a cue to indicate global or local T1.