Add warmup trials to each experimental block
Runs a set of warmup trials at the beginning of experimental, but not
practice, blocks. Add an additional loop to allow for these trials. Set up
logic to quit after the appropriate number of trials have been run, and
also add logic to quit after the desired number of practice trials on
practice blocks.