Switch to a tight loop for RSVP stream
Instead of waiting until just before the next time, have a tight loop that
checks regularly. That allows me to set up the next time based on the last
flip instead of after the end of a bunch of drawing. Add a clock and
extract the ISI and stimulus durations into separate variables to reduce a
tiny bit of overhead.