Reduce t2 lags
Currently we're running at frames of 150 ms each, so the previous lags of 3
(450 ms) and 7 (1050 ms) were towards the end of the AB and unnecessarily
far after it. This change allows us to test T2 during the early/middle
phase of the AB, and reduce the frame length and possibly the total time of
the experiment (at least modestly).