Move data saving step earlier
We have been experiencing intermittent crashes on Windows-based computers
(10 and 11). I think it might be due to someone clicking the mouse during
the experiment, though I'm not sure. In some cases this has happened at the
very end while displaying the accuracy percentages, and the crash means no
data gets saved at all.

Move the data save step to before the final screen so if the crash occurs
then, at least the data has been written to a file already.