Anonymize subject IDs in a more consistent way
Sort the subjects by run day/time so they are in a consistent order, then
set a constant seed so the new IDs get randomized in the same order. That
means that a subject should always receive the same ID, though the mapping
isn't actually stored anywhere.

This is kinda starting to feel silly...