Add separate cleaning and checking file
Takes the code that used to be in 00-GatherData.Rmd to clean, check, and
aggregate the data, and puts it into a separate file so we can develop and
test this code separately from the time consuming data loading code.

Output to a file with a name based on this one to make it hopefully easier
to pass files between stages of analysis. Also adjust and update some of
the text.

Remove the file that extracted the data from Fall 2023 and put it into the
file that handles that analysis. Seems more modular that way?