Update for 2AFC for both targets
Revise code to pick from a subset of stimuli for each of T1 and T2, making
sure that those letters aren't part of the distractor set, and allowing
them to be a different color from the distractors. Adjust response
collection to compensate. Also adjust settings to account for this design
and for the different set of hierarchical letters and stimulus sizes from
HierPy.