Revamp color handling
Have every visual object except ImageStim define its color space based on
a setting defined at the top of the file, and also base the main window's
background color on a setting. This makes it easier to centralize colors
and color space. Also change the background color to a lighter gray to
match the stimuli, for better contrast with both white and black stimuli.

For some reason, PsychoPy visual objects each have their own color space
instead of inheriting it from the main window, and that color space
defaults to -1:1. My preference is to have it be 0:255, but then that has
to be set separately for each object.