Separate out data filtering in RSVPLG01 and 02 analyses
Create a separate Rmd file that just has optional filtering commands so
that all other analysis files can operate on a common dataset. For
RSVPLG01 this means pulling the filtering code from the 05-Exploratory
file. For both experiments it means returning to loading the binary RData
file generated by the filtering step, which doesn't need to be included in
the git repo.